Created excavator views, urls, excavator.html and updated admin pannel view. Tested connection and if data display on site properly.

+++ excavator/models.py
@@ -5,7 +5,7 @@ class Excavator(models.Model):
 
     name = models.CharField(max_length=254)
     model = models.CharField(max_length=254)
-    production_year = models.DecimalField(max_digits=6, decimal_places=2)
+    production_year = models.IntegerField()
     weight = models.DecimalField(max_digits=6, decimal_places=2)
     description = models.TextField()
     price = models.DecimalField(max_digits=6, decimal_places=2)
